+++ resources/ReleaseNotes.md
@@ -10,9 +10,10 @@
 ## Other
 
 - Add variable continuity definition for science variables in `timetable`
+- Add shutdown event as final event in `mag.Science` event table
+- Add property `Harness` in `mag.meta.Science` to describe sensor harness
 - Simplify how columns are filtered in loading science and I-ALiRT
 - Make `mag.process.Calibration` more flexible to custom variable names
-- Add property `Harness` in `mag.meta.Science` to describe sensor harness
 - Allow plotting more than one event in `mag.graphics.view.Field`
 - Change event type to `categorical` when data is not numeric
 - Add option to create folder in `mag.graphics.savePlots`